pyproject.toml
[project]
name = "playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "einops>=0.8.1",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "huggingface-hub>=0.36.0",
    "ipykernel>=6.31.0",
    "jupyterlab>=4.5.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.79",
    "langgraph>=0.6.11",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "openai>=2.8.1",
    "orjson>=3.11.4",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "plotly>=6.5.0",
    "pyarrow>=21.0.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]
